requirements.txt
hydra-core==1.3.2
librosa==0.11.0
numpy==2.3.0
pydub==0.25.1
scikit_learn==1.7.0
scipy==1.15.3
soundfile==0.13.1
torch==2.6.0
torchaudio==2.2.2+cu121
torchvision==0.21.0
tqdm==4.67.1
utils==1.0.2
